+++ src/main/java/finalproject/leejeonmoon/domain/dto/CreateAccessTokenResponseDto.java
@@ -1,4 +1,4 @@
-package finalproject.leejeonmoon.domain.member.dto;
+package finalproject.leejeonmoon.domain.dto;
 
 public record CreateAccessTokenResponseDto(
         String accessToken) {
