+++ src/main/java/finalproject/leejeonmoon/domain/dto/ScheduledService.java
@@ -1,10 +1,10 @@
-package finalproject.leejeonmoon.domain.member.dto;
+package finalproject.leejeonmoon.domain.dto;
 
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.scheduling.annotation.Scheduled;
 import org.springframework.stereotype.Component;
-import finalproject.leejeonmoon.domain.member.service.NotificationService;
+import finalproject.leejeonmoon.domain.service.NotificationService;
 
 import java.util.concurrent.ExecutionException;
