+++ src/main/java/finalproject/leejeonmoon/domain/entity/Notification.java
@@ -1,16 +1,7 @@
-package finalproject.leejeonmoon.domain.member.entity;
+package finalproject.leejeonmoon.domain.entity;
 
 import jakarta.persistence.*;
-import jakarta.validation.constraints.Email;
-import jakarta.validation.constraints.NotBlank;
-import jakarta.validation.constraints.NotNull;
 import lombok.*;
-import org.springframework.security.core.GrantedAuthority;
-import org.springframework.security.core.authority.SimpleGrantedAuthority;
-import org.springframework.security.core.userdetails.UserDetails;
-
-import java.util.Collection;
-import java.util.List;
 
 @Entity
 @Getter
