+++ src/main/java/finalproject/leejeonmoon/domain/repository/AlarmRepository.java
@@ -1,0 +1,12 @@
+package finalproject.leejeonmoon.domain.repository;
+
+
+import finalproject.leejeonmoon.domain.entity.Alarm;
+import finalproject.leejeonmoon.domain.entity.Member;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import java.util.List;
+
+public interface AlarmRepository extends JpaRepository<Alarm, Long> {
+    List<Alarm> findByMember(Member member);
+}
