+++ src/main/java/finalproject/leejeonmoon/domain/repository/MemberRepository.java
@@ -1,6 +1,6 @@
-package finalproject.leejeonmoon.domain.member.repository;
+package finalproject.leejeonmoon.domain.repository;
 
-import finalproject.leejeonmoon.domain.member.entity.Member;
+import finalproject.leejeonmoon.domain.entity.Member;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
