+++ src/main/java/finalproject/leejeonmoon/domain/repository/NotificationRepository.java
@@ -1,13 +1,12 @@
-package finalproject.leejeonmoon.domain.member.repository;
+package finalproject.leejeonmoon.domain.repository;
 
-import finalproject.leejeonmoon.domain.member.entity.Member;
-import finalproject.leejeonmoon.domain.member.entity.Notification;
+import finalproject.leejeonmoon.domain.entity.Member;
+import finalproject.leejeonmoon.domain.entity.Notification;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
 public interface NotificationRepository extends JpaRepository<Notification, Long>  {
     Optional<Object> findByMember(Member member);
-
     boolean existsByToken(String token);
 }
