+++ src/main/java/finalproject/leejeonmoon/domain/service/FCMInitializer.java
@@ -1,4 +1,4 @@
-package finalproject.leejeonmoon.domain.member.service;
+package finalproject.leejeonmoon.domain.service;
 
 import com.google.auth.oauth2.GoogleCredentials;
 import com.google.firebase.FirebaseApp;
