+++ src/main/java/finalproject/leejeonmoon/global/config/jwt/TokenProvider.java
@@ -5,7 +5,7 @@ import io.jsonwebtoken.Header;
 import io.jsonwebtoken.Jwts;
 import io.jsonwebtoken.SignatureAlgorithm;
 import lombok.RequiredArgsConstructor;
-import finalproject.leejeonmoon.domain.member.entity.Member;
+import finalproject.leejeonmoon.domain.entity.Member;
 import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
 import org.springframework.security.core.Authentication;
 import org.springframework.security.core.authority.SimpleGrantedAuthority;
